requirements.txt
telebot
requests
json
python-dotenv
